Put the presave hook next to the model for now
